Improve virsh autocompletion (base framework)
Previously, virsh was able to complete initial command names,
as well as the names of flag options.  However, this completion
was clunky for a number of reasons:

- it would provide a flag as an option for completion
  even if it had already been used

- it did not support completion of positional arguments

- it would fall back to filename completion, even when
  a file was not an appropriate argument

This commit improves virsh autocompletion by actually parsing
the line using the extracted parsing logic mentioned in the previous
commit.  This allows for proper completion of positional arguments
and flag arguments, as well as preventing a flag from being passed
multiple times.

Additionally, it removes the default behavior of falling back to filename
completion when no matches are found.  Now, the VSH_COMPLETE_AS_FILE
completer_flag may be passed with a null completer function to use
this functionality.  Otherwise, no completion options are provided.

This commit also introduces a helper macro called VSH_STRING_COMPLETER
to aid in the writing of completers which have a fixed set of strings.
A call of `VSH_STRING_COMPLETER(ctl, SomeOption, "string1", "string2")`
will generate a function `vshCompleteSomeOption`, suitable for use
as a custom completer.  The macro may accept up to 63 different strings
(a limitation of C VA_ARGS macros).  Additionally, definition of
vshStrndup (which calls virStrndup) was introduced, following the
vsh-prefixed memory-related functions.

Furthermore, two new commands have been introduced:

- 'complete' takes either a quoted partial line, or
  and unquoted partial line following a '--'.  It returns
  a newline-separated list of potential completions as expected
  by readline.  It is useful for writing commandline completers
  (bash completion, etc) and for testing.

- 'fake-command' takes a series of flags and data arguments, and
  prints those arguments which are present.  It is used to test
  the completion in cases not possible to cover with 'echo'.

Finally, a suite a tests have been introduced.