Improve virsh autocompletion (domain completer)
This patch introduces a custom completer for domains,
and sets it to be used for all of the "domain" arguments.

In order to facilitate this, the functions involved in
retrieve in a list of domains were moved from
virsh-domain-monitor to virsh-completer, where they
can be included by any file that needs to use them.