Improve virsh autocompletion (global ctl object)
This patch introduces a way for completers to retrieve the
current vshControl object by using the vshGetCompleterCtl()
macro.  When readline is enabled, the main method stores the
vshControl pointer in a variable that is file-global to virsh.c.
Then, that pointer can be retrieved by the _vshGetCompleterCtl()
function, which is mapped to by the macro.  If readline is not
enabled, the macro simply maps to NULL.