Improve virsh autocompletion (extract parser)
This commit extracts the parsing logic from vshCommandParse
so that it can be used by other methods.  The vshCommandParse
method is designed to parse full commands and error out on
unknown information, so it is not suitable to simply use it
for autocompletion.  Instead, the logic has been extracted.

The new method essentially performs one pass of the loop previously
done by vshCommandParse.  Depending on what happens, instead of erroring
or continuing the loop, it returns a value from an enum indicating the
current state.  It also modifies several arguments as appropriate.

Then, a caller can choose to deal with the state, or may simply ignore
the state when convenient (vshCommandParse deals with the state,
so as to continue providing its previous functionality, while a
completer could choose to ignore states involving unknown options,
for example).